+++ server/src/main/java/com/linbit/linstor/api/pojo/VlmDfnWithCreationPayloadPojo.java
@@ -3,6 +3,8 @@ package com.linbit.linstor.api.pojo;
 import com.linbit.linstor.core.apis.VolumeDefinitionApi;
 import com.linbit.linstor.core.apis.VolumeDefinitionWithCreationPayload;
 
+import javax.annotation.Nullable;
+
 public class VlmDfnWithCreationPayloadPojo implements VolumeDefinitionWithCreationPayload
 {
     private final VolumeDefinitionApi vlmDfnApi;
@@ -26,4 +28,12 @@ public class VlmDfnWithCreationPayloadPojo implements VolumeDefinitionWithCreati
         return drbdMinorNr;
     }
 
+    /**
+     * Add this point the passphrase is already stored in the layer data, so always return null.
+     */
+    @Override
+    public @Nullable String passphrase()
+    {
+        return null;
+    }
 }