+++ server/src/main/java/com/linbit/linstor/core/apis/VolumeDefinitionWithCreationPayload.java
@@ -1,7 +1,11 @@
 package com.linbit.linstor.core.apis;
 
+import javax.annotation.Nullable;
+
 public interface VolumeDefinitionWithCreationPayload
 {
     VolumeDefinitionApi getVlmDfn();
     Integer getDrbdMinorNr();
+
+    @Nullable String passphrase();
 }